% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{allCombApply}
\alias{allCombApply}
\title{apply a function to all combination of columns}
\usage{
allCombApply(inmat, FUN, ...)
}
\value{
data.frame (or vector) with (n*(n-1)/2) columns (or elements)
}
\description{
\code{allCombApply} applies a function to all combination of columns
}
\details{
\code{FUN} should have two (vector) parameters such as \code{function(v1, v2) {cor(v1, v2)}}
}
\examples{
allCombApply(matrix(c(1, 2, 3, 4, 5, 6, 7, 8), nrow = 2), function(v1, v2) {v1-v2})
}
