% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{allCombApplyM}
\alias{allCombApplyM}
\title{applies given function to all combinations of matrix columns}
\usage{
allCombApplyM(inmat, FUN, ...)
}
\value{
symmetric matrix
}
\description{
applies given function to all combinations of matrix columns
}
\details{
\code{FUN} should have two (vector) parameters such as \code{function(v1, v2) {cor(v1, v2)}}
}
\examples{
allCombApplyM(matrix(rnorm(25), nrow = 5), FUN = cor)
}

