% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{amerge}
\alias{amerge}
\title{Merge two datasets based on both exact and approximate keys}
\usage{
amerge(x, y, aby, method = "jw", tol = 0.1, p = 0.1, ...)
}
\arguments{
\item{x}{A data.frame or data.table.}

\item{y}{A data.frame or data.table.}

\item{aby}{name of a string key for approximate matching}

\item{method}{Method for distance calculation.}

\item{tol}{tolerance of dissimilarity between string keys in matched datasets}

\item{p}{see \code{stringdist} package}

\item{...}{other parameters passed to \code{\link{merge}}}
}
\description{
\code{amerge} merges two datasets by exact keys and filter out rows that do not approximately match by given string keys.
}
\examples{

dt1 <- data.table(id   = c(1,      2,             3,      4,                  5),
                  name = c("Aple", "Google inc.", "ABB",  "Sun",              "Sony"),
                  value = rnorm(5))

dt2 <- data.table(id   = c(1,       2,             3,     4,                  5),
                  name = c("Apple", "Google",     "BBC",  "Sun Microsystems", "Sony Corporation"),
                  value = rnorm(5))

amerge(dt1, dt2, by = "id", aby = "name", tol = 0.3)
}
\seealso{
\code{\link{merge}}, \code{\link{stringdist}} and \code{\link{stringsim}}
}
