% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{as.Date2}
\alias{as.Date2}
\title{alias for as.Date function for converting dates from dd.mm.yyyy format}
\usage{
as.Date2(x, ...)
}
\arguments{
\item{x}{A string vector containing date encoded as dd.mm.yyyy}
}
\description{
alias for as.Date function for converting dates from dd.mm.yyyy format
}
