% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{as_ascii}
\alias{as_ascii}
\alias{clDiacr}
\title{convert to ASCII}
\usage{
as_ascii(x, from = "", to = "ASCII//TRANSLIT", ...)

clDiacr(x, from = "", to = "ASCII//TRANSLIT", ...)
}
\description{
convert to ASCII
}
