% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{askSave}
\alias{askSave}
\title{Save file after asking the user for confirmation}
\usage{
askSave(data, filename = substitute(data), filepath = "../output/",
  csv = 0, row.names = F, ...)
}
\arguments{
\item{data}{data file to be writen}

\item{filename}{name of the file, defaults to name of the object}

\item{filepath}{path where to save the file, defaults to "../output/"}

\item{csv}{write the data into a csv2 as well?}
}
\description{
Save file after asking the user for confirmation
}
