% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{askUser}
\alias{askUser}
\title{ask for user's input}
\usage{
askUser(question, answers = c("y", "n"), default = 1)
}
\arguments{
\item{\code{question}}{question to be prompted}

\item{\code{answers}}{vector of possible answers}

\item{\code{default}}{index of default answer}
}
\value{
valid answer, i.e. answer is in \code{answers}
}
\description{
ask for user's input
}
\details{
In case \emph{default} is in range of possible answers, the default answer is uppercased and empty input (return key) is treated as the default answer.
}

