% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{clDiacr}
\alias{clDiacr}
\title{convert special letters to some normal equivalent characters}
\usage{
clDiacr(x, from = "", to = "ASCII//TRANSLIT", ...)
}
\description{
convert special letters to some normal equivalent characters
}
