% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{clSpecialChars}
\alias{clSpecialChars}
\title{cleans strings for use as a file name -> for saving charts with names based on chart titles}
\usage{
clSpecialChars(x)
}
\description{
cleans strings for use as a file name -> for saving charts with names based on chart titles
}
\details{
replaces spaces, special chars and and end-of-line charactgers by underscore characters
}

