% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{\%pp\%}
\alias{\%pp\%}
\alias{\%p\%}
\alias{\%pc\%}
\title{Concatenation operators}
\usage{
a \%pp\% b

a \%p\% b

x \%pc\% c
}
\description{
Concatenation operators
}
\details{
\describe{
  \item{\%pp\%}{\code{a \%pp\% b} corresponds to \code{paste(a, b)}}
  \item{\%p\%}{\code{a \%p\% b} corresponds to \code{paste0(a, b)}}
  \item{\%pc\%}{\code{x \%pc\% c} corresponds to \code{paste(x, collapse = c)}}
}
}
