% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{convertClass}
\alias{convertClass}
\title{convert a vector to a specified type}
\usage{
convertClass(v, classtype)
}
\arguments{
\item{\code{v}}{input vector}

\item{\code{classtype}}{"character", "numeric", "logical", "integer" or "double".}
}
\value{
vector of the specified type
}
\description{
convert a vector to a specified type
}
