% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{convertClasses}
\alias{convertClasses}
\title{convert columns in a data.frame to a specified types}
\usage{
convertClasses(df, classes)
}
\arguments{
\item{\code{df}}{input data.frame that should be converted}

\item{\code{classtype}}{vector of target types : "character", "numeric", "logical", "integer" or "double".}
}
\value{
data.frame
}
\description{
convert columns in a data.frame to a specified types
}

