% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leis.R
\name{convertEntityList2df}
\alias{convertEntityList2df}
\title{Transform given list of legal entities from \code{scrapeLEI} funtion to a data.frame}
\usage{
convertEntityList2df(lEntities, keepDetails = c("code", "legalCountry",
  "hqCountry", "legalName", "registrysIdentifier", "otherLegalName"),
  wide = T)
}
\arguments{
\item{\code{lEntities}}{list of legal entities as returned by \code{scrapeLEI}.}

\item{\code{wide}}{logical; set \code{TRUE} if you want the data.frame to be reshaped from long to wide.}
}
\value{
data.frame
}
\description{
Transform given list of legal entities from \code{scrapeLEI} funtion to a data.frame
}

