% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ctrlP}
\alias{ctrlP}
\title{read from clipboard for easy copy-paste from spreasdsheet (Excel)
\code{ctrlP} reads clipboard (copied from excel or other spreasheet) and returns a \code{data.frame}.}
\usage{
ctrlP(dec = ",", sep = "\\t", rn = NULL, ...)
}
\description{
read from clipboard for easy copy-paste from spreasdsheet (Excel)
\code{ctrlP} reads clipboard (copied from excel or other spreasheet) and returns a \code{data.frame}.
}
