% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{diff_fill}
\alias{diff_fill}
\title{alternative version of diff() returning constant-length vector}
\usage{
diff_fill(x, lag = 1, differences = 1, fill = NA)
}
\arguments{
\item{x}{numeric/integer vector (see help for diff())}

\item{x}{lag (see help for diff())}

\item{x}{differences (see help for diff())}

\item{x}{fill value to fill in missing values}
}
\value{
vector of the same length and as \code{x}
}
\description{
Wrapper to diff() function that fill in values at the beginning of the resulting vector
}
\examples{
str(diff(1:10))
str(diff_fill(1:10))
str(diff_fill(1:10, fill = NULL))
}
