% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{dropAllNA}
\alias{dropAllNA}
\title{Takes an array and return a copy without cols/rows that are NAs-only}
\usage{
dropAllNA(data, mar = 1, id = NULL)
}
\arguments{
\item{\code{data}}{matrix or data.frame}

\item{\code{mar}}{margin (1 for rows, 2 for columns)}

\item{\code{id}}{a character vecor with column names that are id/keys and should by skipped when testing for missingness}
}
\value{
an array of the same type as the input \code{data} object.
}
\description{
Takes an array and return a copy without cols/rows that are NAs-only
}
\examples{
tmp <- data.frame(a = 1:5, b = c(34, 45, 56, NA, NA), c = c(2, 3, 4, 5, NA))
deleteAllNA(tmp)
deleteAllNA(tmp, id = "a")
deleteAllNA(tmp, id = c("a", "b"))
}

