% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{dropNA}
\alias{dropNA}
\title{return a copy of a vector omitting NA values}
\usage{
dropNA(vect)
}
\arguments{
\item{\code{vect}}{a vector of any type.}
}
\value{
a vector of the same type as \code{vect}.
}
\description{
return a copy of a vector omitting NA values
}

