% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{fillNAs}
\alias{fillNAs}
\title{fill missing values by constant}
\usage{
fillNAs(x, by_val = 0)
}
\arguments{
\item{\code{x}}{input vector}

\item{\code{by_val}}{substitute value}
}
\description{
fill missing values by constant
}

