% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{group_str}
\alias{group_str}
\title{Collapse character vector (strings) into groups.}
\usage{
group_str(x, l, pass = TRUE, USE.NAMES = TRUE)
}
\arguments{
\item{x}{a character vector}

\item{l}{a (usually named) list of character vectors
#' @export}
}
\description{
For each string in a vector \code{group_str} finds first match in list of vectors and returns name of the vector.
}

