% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{group_vec}
\alias{group_vec}
\title{Collapse vector into groups.}
\usage{
group_vec(x, l, no.match = NULL, USE.NAMES = TRUE)
}
\arguments{
\item{x}{a character vector}

\item{l}{a (usually named) list of character vectors}

\item{no.match}{value that should be returned if no match is found. Usually NA or "other". If NULL (default) the original value itself is returned.}

\item{USE.NAMES}{logical passed to sapply indicating whether a result should be a named vector (for character input vectors only).}
}
\description{
For each value in a vector \code{group_vec} finds first match in list of vectors and returns name of the vector.
}

