% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{hist4Calc}
\alias{hist4Calc}
\title{read a series from (excel, Calc, ...) clipboard, draws its histogram and save results to clipbaord (for excel, Calc, ...)}
\usage{
hist4Calc(...)
}
\arguments{
\item{\code{...}}{parameters tp be passed to \code{ctrlC} and \code{ctrlP} functions}
}
\description{
read a series from (excel, Calc, ...) clipboard, draws its histogram and save results to clipbaord (for excel, Calc, ...)
}
