% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{iconv.data.table}
\alias{iconv.data.table}
\title{apply iconv on all character columns of a copy of a data.table}
\usage{
iconv.data.table(DT, ...)
}
\arguments{
\item{...}{args are passed to iconv}
}
\description{
apply iconv on all character columns of a copy of a data.table
}
