% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{iconvFile}
\alias{iconvFile}
\title{convert file encoding}
\usage{
iconvFile(from_file, to_file, from_encoding = "windows-1250",
  to_encoding = "UTF-8")
}
\arguments{
\item{from_file}{source file path}

\item{to_file}{destination file path}

\item{from_encoding}{source endoding, Default: 'windows-1250'}

\item{to_encoding}{target encoding, Default: 'UTF-8'}
}
\value{
TRUE if succesfull
}
\description{
convert a file from one encoding to another
}
\details{
convert a file from one encoding to another
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
