% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{install_github_proxy}
\alias{install_github_proxy}
\title{set proxy seetings and call install_github()}
\usage{
install_github_proxy(..., proxy = getIEProxy()[1],
  port = as.integer(getIEProxy()[2]))
}
\description{
set proxy seetings and call install_github()
}

