% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{mergeAll}
\alias{mergeAll}
\title{merges multiple data.frames or data.tables at once}
\usage{
mergeAll(datasets, all = T, ...)
}
\arguments{
\item{datasets}{is a list of data.frames or data.tables...}
}
\value{
a single data.frame/data.table
}
\description{
\code{mergeAll} merges multiple data.frames with the same id vars
}
