% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{\%+n\%}
\alias{\%*n\%}
\alias{\%+n\%}
\alias{\%-n\%}
\alias{\%/n\%}
\title{Addition of two numbers/vectors/matrices while replacing all NAs by zeros.}
\usage{
a \%+n\% b

a \%-n\% b

a \%*n\% b

a \%/n\% b
}
\description{
Addition of two numbers/vectors/matrices while replacing all NAs by zeros.

Deduct two numbers/vectors/matrices while replacing all NAs by zeros.

Product of two numbers/vectors/matrices while replacing all NAs by zeros.

Division of two numbers/vectors/matrices while replacing all NAs by zeros.
}
\details{
the two objects of appropriate class types.
}
\examples{
c(1, 2, 4) \%+n\% c(3, 6, NA)
c(1, 2, 4) \%-n\% c(3, 6, NA)
c(1, 2, NA) \%*n\% 3
c(3, 6, NA) \%/n\% 3
}

