% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{normalize}
\alias{normalize}
\alias{rescale}
\title{rescale numeric objects}
\usage{
normalize(v, ...)

rescale(x, low = 0, high = 1, ...)
}
\arguments{
\item{\code{v}}{vector/matrix to be rescaled}

\item{\code{low}}{minimum of the resulting interval}

\item{\code{high}}{maximum of the resulting interval}

\item{\code{...}}{parameters passed to \code{min} and \code{max} (or \code{range}) methods}
}
\value{
\code{normalize} returns vector/matrix rescaled to the \code{[0;1]} interval

\code{rescale} returns vector/matrix rescaled to the \code{[low;high]} interval
}
\description{
rescale numeric objects
}
