% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{omitNALines}
\alias{omitNALines}
\title{deletes all-NAs rows and columns (deprecated, see dropAllNA)}
\usage{
omitNALines(data, mar = c(1, 2), id = NULL)
}
\description{
deletes all-NAs rows and columns (deprecated, see dropAllNA)
}
