% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eiu.R
\name{readEIU}
\alias{readEIU}
\title{read xlsx file as downloaded from EIU and return panel data in the form of (country + year ~ variable)}
\usage{
readEIU(file, dfRenameSeries = NULL, ...)
}
\arguments{
\item{\code{file}}{file name (path).}

\item{\code{dfRenameSeries}}{optional.}
}
\description{
read xlsx file as downloaded from EIU and return panel data in the form of (country + year ~ variable)
}
\details{
depends on data.table and xlsx packages
}

