% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{recodeStrings}
\alias{recodeStrings}
\title{recode vector of strings using a convertion table}
\usage{
recodeStrings(input, conversionTable, from = 1, to = 2, replaceNA = T)
}
\arguments{
\item{\code{input}}{input vector (usually strings)}

\item{\code{conversionTable}}{data frame of at least two (from and to) columns}

\item{\code{from, to}}{indeces of from/to columns}

\item{\code{replaceNA}}{if TRUE all values that couldn't be matched will remain the origianl values, otherwise left as NAs}
}
\description{
recode vector of strings using a convertion table
}
\details{
...
}

