% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{saveggplot}
\alias{saveggplot}
\title{save last ggplot to a file}
\usage{
saveggplot(file = clDiacr(last_plot()$labels$title), dir = "../output/",
  format = ".wmf", width = 9, height = 9)
}
\arguments{
\item{file}{- name of the file. By default title of the plot is used.}

\item{dir}{- destination folder - "..\/output\/" by default.}

\item{format}{- ".wmf" or ".pdf"}

\item{width}{- width of the plot}

\item{height}{- height of the plot}
}
\description{
\code{saveggplot} save last printed ggplot to a file
}

