% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leis.R
\name{scrapeLEI}
\alias{scrapeLEI}
\title{Scrape attributes of entitities from lei-lookup.com given list of LEI codes}
\usage{
scrapeLEI(lei_codes, token = "", proxy = paste(getIEProxy(), collapse =
  ":"), showMissing = F)
}
\arguments{
\item{\code{lei_codes}}{character verctor of LEI codes}

\item{\code{token}}{token for access to lei-lookup.com REST API}

\item{\code{proxy}}{proxy settings}
}
\value{
list of legal entities with all attributes
}
\description{
Scrape attributes of entitities from lei-lookup.com given list of LEI codes
}
\examples{
lei_vec <- c("259400DZXF7UJKK2AY35", "some misspelled LEI", "529900LN3S50JPU47S06", "some other LEI")
scrapeLEI(lei_vec)
scrapeLEI(lei_vec, showMissing = T)

convertEntityList2df(scrapeLEI(lei_vec))
convertEntityList2df(scrapeLEI(lei_vec), wide = F)
}
