% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{seqv}
\alias{seqv}
\title{seq function on vectors}
\usage{
seqv(from, to, ..., simplify = F)
}
\arguments{
\item{\code{from, to}}{the starting and (maximal) end values vectors of the sequences.}

\item{\code{simplify}}{logical: whether the function returns a matrix/list (depending on reqularity of \code{from-to}) (\code{TRUE}) a vector (\code{FALSE}) .}
}
\description{
seq function on vectors
}
\details{
If the \code{from - to} vector is not constant vector and \code{simplify = F}, \code{seqv} returns a list.
}

