% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{iconv_set.data.table}
\alias{iconv_set.data.table}
\alias{iconv_set}
\alias{seticonv}
\title{apply iconv on all character columns of a data.table}
\usage{
\method{iconv_set}{data.table}(DT, ...)

iconv_set(obj, ...)

seticonv(obj, ...)
}
\arguments{
\item{...}{args are passed to iconv}
}
\description{
apply iconv on all character columns of a data.table
}
