% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{shortenString}
\alias{shortenString}
\title{shorten strings by cutting characters from the middle}
\usage{
shortenString(x, max)
}
\arguments{
\item{x}{character object to be shortened}

\item{max}{maximum number of letters of the resulting characters}
}
\description{
shorten strings by cutting characters from the middle
}
