% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{source_search}
\alias{source_search}
\title{Search for a file in different places and source the first hit.}
\usage{
source_search(file, places, ...)
}
\arguments{
\item{\code{file}}{name of the file}

\item{\code{places}}{character vector of places (directories) where to search for the file}

\item{\code{...}}{other arguments will be passed to \code{source}}
}
\description{
Search for a file in different places and source the first hit.
}
