% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{split2Batches}
\alias{split2Batches}
\title{split vector into batches}
\usage{
split2Batches(x, size)
}
\arguments{
\item{x}{vector}

\item{size}{(maximum) size of a single batch}
}
\description{
split vector into batches
}

