% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_shrt}
\alias{str_shrt}
\title{shorten strings by cutting characters from the middle - alias for shortenString()}
\usage{
str_shrt(x, max)
}
\description{
shorten strings by cutting characters from the middle - alias for shortenString()
}
