% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VHtools.R
\docType{package}
\name{VHtools-package}
\alias{VHtools}
\alias{VHtools-package}
\title{This paclage provides various help functions for daily work with strings, clipboard etc...}
\description{
This paclage provides various help functions for daily work with strings, clipboard etc...
}
