% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{winply_v}
\alias{winply_v}
\title{Apply function to rolling/moving windows of changing width}
\usage{
winply_v(x, fun, win, fill = NA, ...)
}
\description{
Apply function to rolling/moving windows of changing width
}
