% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{diff_fill}
\alias{diff_fill}
\title{An alternative version of diff() returning vector of the same length as input}
\usage{
diff_fill(x, lag = 1L, differences = 1L, fill = NA)
}
\arguments{
\item{x}{numeric/integer vector (see help for diff())}

\item{lag}{(see help for diff())}

\item{differences}{(see help for diff())}

\item{fill}{value to fill in missing values}
}
\value{
vector of the same length and as \code{x}
}
\description{
Wrapper to diff() function that fills in user-defined values (e.g. NA) at the beginning of the resulting vector
}
\examples{
str(diff(1:10))
str(diff_fill(1:10))
str(diff_fill(1:10, fill = NULL))
}
