% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{instgithubproxy}
\alias{instgithubproxy}
\title{set proxy seetings and call install_github()}
\usage{
instgithubproxy(..., proxy = getIEProxy()[1],
  port = as.integer(getIEProxy()[2]))
}
\description{
set proxy seetings and call install_github()
}

