% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{normalize}
\alias{normalize}
\title{returns normalized vector (matrix)}
\usage{
normalize(v)
}
\arguments{
\item{\code{v}}{vector/matrix to be normalized}
}
\description{
returns normalized vector (matrix)
}
